🐛 Fixed a bug where the node ID was not generated correctly when creating a new node in a new workflow, accompanied by a crypto.UUID error.
🐛 Fixed a bug where the node ID was not generated correctly when creating a new node in a new workflow, accompanied by a crypto.UUID error.
