+++ src/services/NodeClipboardService.ts
@@ -12,16 +12,10 @@ export interface CopiedNode {
 const STORAGE_KEY = 'comfy_mobile_node_clipboard';
 const MAX_COPIES = 3;
 
-// Fallback UUID generator for non-secure contexts
-function generateUUID() {
-    if (typeof crypto !== 'undefined' && crypto.randomUUID) {
-        return crypto.randomUUID();
-    }
-    return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
-        var r = Math.random() * 16 | 0, v = c == 'x' ? r : (r & 0x3 | 0x8);
-        return v.toString(16);
-    });
-}
+import { generateUUID } from '@/utils/uuid';
+
+// Removed internal generateUUID function
+
 
 export const NodeClipboardService = {
     saveNode: (node: Omit<CopiedNode, 'id' | 'timestamp'>) => {
