+++ src/utils/uuid.ts
@@ -1,0 +1,13 @@
+/**
+ * Generate a UUID (Version 4)
+ * Uses crypto.randomUUID() if available, otherwise falls back to a math-based generator.
+ */
+export function generateUUID(): string {
+    if (typeof crypto !== 'undefined' && crypto.randomUUID) {
+        return crypto.randomUUID();
+    }
+    return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
+        var r = Math.random() * 16 | 0, v = c == 'x' ? r : (r & 0x3 | 0x8);
+        return v.toString(16);
+    });
+}
